refactor(ui): extract EntityListPage from Resources/Users/Identities pages
Creates a shared EntityListPage scaffold that owns all repeated JSX
(header, tag bar, create-tag form, filter bar, action bar, table,
pagination). Each of the three list pages shrinks to a config wrapper
passing title, endpoints, column definitions, and row-render callbacks.
Eliminates ~700 lines of near-identical JSX.
