+++ app/ui/src/components/UsersPage.jsx
@@ -1,9 +1,6 @@
-﻿import { useMemo, useState, useEffect } from 'react';
-import { useAuth } from '@ui/auth/AuthGate';
-import useEntityPage from '@ui/hooks/useEntityPage';
-import FilterBar from './FilterBar';
+import { useMemo, useState, useEffect } from 'react';
+import EntityListPage from './EntityListPage';
 import DeletedBadge from './DeletedBadge';
-import { TAG_COLORS } from '@ui/utils/colors';
 
 const FIELD_LABELS = {
   department: 'Department',
@@ -32,8 +29,6 @@ const TABLE_COLUMNS = [
 // Sub-tabs for principalType. The Principals table is a universal identity
 // store, so the "Users" page now lists more than just humans — splitting it
 // by type keeps each view manageable when SP/MI/AIAgent sync is enabled.
-// The tab label is what's shown; the value is matched against the column.
-// 'all' is a sentinel for "no filter".
 const PRINCIPAL_TYPE_TABS = [
   { key: 'all',              label: 'All' },
   { key: 'User',             label: 'Users' },
@@ -64,343 +59,73 @@ function writeTypeToHash(tab) {
 }
 
 export default function UsersPage({ onOpenDetail }) {
-  const { authFetch } = useAuth();
   const [activeTypeTab, setActiveTypeTab] = useState(readTypeFromHash);
-
   useEffect(() => { writeTypeToHash(activeTypeTab); }, [activeTypeTab]);
 
-  // Memoise so useEntityPage's filtersObj memo isn't busted every render.
+  // Memoised so useEntityPage's filtersObj memo isn't busted every render.
   const baseFilters = useMemo(
     () => (activeTypeTab === 'all' ? null : { principalType: activeTypeTab }),
     [activeTypeTab],
   );
 
-  const ep = useEntityPage({
-    authFetch,
-    entityType: 'user',
-    listEndpoint: '/api/users',
-    columnsEndpoint: '/api/user-columns-page',
-    tagFilterKey: '__userTag',
-    baseFilters,
-  });
+  // Hide `principalType` from Filters when a specific sub-tab is active —
+  // the tab is the authoritative selector, so two controls for the same value
+  // would be confusing.
+  const customizeFilterFields = useMemo(
+    () => activeTypeTab === 'all'
+      ? null
+      : (fields) => fields.filter(f => f.key !== 'principalType'),
+    [activeTypeTab],
+  );
 
-  // Hide `principalType` from the Filters dropdown only when a specific
-  // sub-tab is active — the tab is the authoritative selector there, and
-  // having both would create two ways to set the same value. On the "All"
-  // tab no type is pinned, so leave `principalType` available as a regular
-  // filter option.
-  const filterFields = useMemo(
-    () => ep.getFilterFields(FIELD_LABELS).filter(f =>
-      !(activeTypeTab !== 'all' && f.key === 'principalType')
-    ),
-    [ep, activeTypeTab],
+  const subTabBar = (
+    <div className="border-b border-gray-200 dark:border-gray-700 mb-4">
+      <nav className="flex gap-1 -mb-px">
+        {PRINCIPAL_TYPE_TABS.map(tab => (
+          <button
+            key={tab.key}
+            onClick={() => setActiveTypeTab(tab.key)}
+            className={`px-4 py-2 text-sm font-medium border-b-2 transition-colors ${
+              activeTypeTab === tab.key
+                ? 'border-indigo-600 text-indigo-700'
+                : 'border-transparent text-gray-500 dark:text-gray-400 hover:text-gray-700 dark:hover:text-gray-200 hover:border-gray-300 dark:hover:border-gray-600'
+            }`}
+          >
+            {tab.label}
+          </button>
+        ))}
+      </nav>
+    </div>
   );
 
   return (
-    <div className="max-w-7xl mx-auto">
-      {/* Header */}
-      <div className="flex items-center gap-4 mb-4">
-        <h2 className="text-lg font-semibold text-gray-900 dark:text-white">Users</h2>
-        <span className="text-sm text-gray-500 dark:text-gray-400">{ep.total.toLocaleString()} total</span>
-      </div>
-
-      {/* Principal-type sub-tabs. Matches the underlined-pills style used by
-          the Admin page's own sub-tab bar so the UX is consistent. */}
-      <div className="border-b border-gray-200 dark:border-gray-700 mb-4">
-        <nav className="flex gap-1 -mb-px">
-          {PRINCIPAL_TYPE_TABS.map(tab => (
-            <button
-              key={tab.key}
-              onClick={() => setActiveTypeTab(tab.key)}
-              className={`px-4 py-2 text-sm font-medium border-b-2 transition-colors ${
-                activeTypeTab === tab.key
-                  ? 'border-indigo-600 text-indigo-700'
-                  : 'border-transparent text-gray-500 dark:text-gray-400 hover:text-gray-700 dark:hover:text-gray-200 hover:border-gray-300 dark:hover:border-gray-600'
-              }`}
-            >
-              {tab.label}
-            </button>
-          ))}
-        </nav>
-      </div>
-
-      {/* Tag management bar */}
-      <div className="flex flex-wrap items-center gap-2 mb-3 text-sm">
-        <span className="font-medium text-gray-600 dark:text-gray-400">Tags:</span>
-        {ep.tags.map(t => (
-          <span
-            key={t.id}
-            className={`inline-flex items-center gap-1 px-2 py-0.5 rounded-full text-xs font-medium cursor-pointer border ${
-              ep.activeTagFilter === t.name
-                ? 'ring-2 ring-offset-1 ring-blue-400'
-                : 'hover:opacity-80'
-            }`}
-            style={{ backgroundColor: t.color + '20', borderColor: t.color, color: t.color }}
-            onClick={() => {
-              if (ep.activeTagFilter === t.name) {
-                ep.removeFilter('__userTag');
-              } else {
-                ep.addFilter('__userTag', t.name);
-              }
-            }}
-            title={`${t.assignmentCount} users tagged — click to filter`}
-          >
-            {t.name}
-            <span className="text-[10px] opacity-70">({t.assignmentCount})</span>
-            <button
-              onClick={(e) => { e.stopPropagation(); ep.deleteTag(t.id); }}
-              className="ml-0.5 hover:opacity-100 opacity-50"
-              title="Delete tag"
-            >
-              &times;
-            </button>
-          </span>
-        ))}
-        <button
-          onClick={() => ep.setShowCreateTag(!ep.showCreateTag)}
-          className="px-2 py-0.5 rounded text-xs text-blue-600 hover:bg-blue-50 dark:hover:bg-blue-900/30 border border-blue-200 dark:border-blue-700 border-dashed"
-        >
-          + New Tag
-        </button>
-      </div>
-
-      {/* Create tag form */}
-      {ep.showCreateTag && (
-        <div className="flex items-center gap-2 mb-3 p-3 bg-gray-50 dark:bg-gray-700/50 border border-gray-200 dark:border-gray-700 rounded-lg text-sm">
-          <input
-            type="text"
-            value={ep.newTagName}
-            onChange={e => ep.setNewTagName(e.target.value)}
-            onKeyDown={e => e.key === 'Enter' && ep.createTag()}
-            placeholder="Tag name..."
-            className="px-2 py-1 border border-gray-300 dark:border-gray-600 rounded text-sm w-48 dark:bg-gray-700 dark:text-gray-200 dark:placeholder-gray-500"
-            autoFocus
-          />
-          <div className="flex items-center gap-1">
-            {TAG_COLORS.map(c => (
-              <button
-                key={c}
-                onClick={() => ep.setNewTagColor(c)}
-                className={`w-5 h-5 rounded-full border-2 ${ep.newTagColor === c ? 'border-gray-800 scale-110' : 'border-transparent'}`}
-                style={{ backgroundColor: c }}
-              />
-            ))}
-          </div>
-          <button
-            onClick={ep.createTag}
-            disabled={!ep.newTagName.trim() || ep.busy}
-            className="px-3 py-1 rounded text-sm font-medium text-white bg-blue-600 hover:bg-blue-700 disabled:opacity-50"
-          >
-            Create
-          </button>
-          <button
-            onClick={() => ep.setShowCreateTag(false)}
-            className="px-2 py-1 rounded text-sm text-gray-500 dark:text-gray-400 hover:bg-gray-200 dark:hover:bg-gray-700"
-          >
-            Cancel
-          </button>
-        </div>
+    <EntityListPage
+      title="Users"
+      entityType="user"
+      listEndpoint="/api/users"
+      columnsEndpoint="/api/user-columns-page"
+      tagFilterKey="__userTag"
+      tableColumns={TABLE_COLUMNS}
+      fieldLabels={FIELD_LABELS}
+      renderEntityCell={(u, openDetail) => (
+        <td className="px-3 py-2 font-medium text-blue-600 hover:text-blue-800 cursor-pointer"
+          onClick={() => openDetail?.('user', u.id, u.displayName)}>
+          {u.displayName}{u.deletedAt && <> <DeletedBadge at={u.deletedAt} /></>}
+        </td>
       )}
-
-      {/* Filter bar + search */}
-      <div className="flex flex-wrap items-center gap-2 mb-3 text-sm">
-        <FilterBar
-          label="Filters:"
-          filterFields={filterFields}
-          activeFilters={ep.activeFilters}
-          getOptionsForField={ep.getOptionsForField}
-          onAddFilter={ep.addFilter}
-          onRemoveFilter={ep.removeFilter}
-          loading={ep.columnsLoading}
-        />
-
-        <div className="border-l border-gray-300 dark:border-gray-600 h-5 mx-1" />
-
-        <input
-          type="text"
-          value={ep.search}
-          onChange={e => ep.setSearch(e.target.value)}
-          placeholder="Search by name or UPN..."
-          className="px-2 py-1 border border-gray-300 dark:border-gray-600 rounded text-xs w-56 dark:bg-gray-700 dark:text-gray-200 dark:placeholder-gray-500"
-        />
-
-        <label className="flex items-center gap-1.5 text-xs text-gray-600 dark:text-gray-400 cursor-pointer select-none" title="Also show users that were deleted in the source system">
-          <input type="checkbox" checked={ep.includeDeleted} onChange={e => ep.setIncludeDeleted(e.target.checked)} />
-          Include deleted
-        </label>
-
-        {ep.hasAnyFilter && (
-          <>
-            <div className="border-l border-gray-300 dark:border-gray-600 h-5 mx-1" />
-            <button
-              onClick={ep.clearAllFilters}
-              className="px-2 py-1 rounded text-xs text-gray-500 dark:text-gray-400 hover:bg-gray-100 dark:hover:bg-gray-700 border border-gray-200 dark:border-gray-700"
-            >
-              Clear all
-            </button>
-          </>
-        )}
-      </div>
-
-      {/* Action bar (visible when items selected) */}
-      {ep.selected.size > 0 && (
-        <div className="flex items-center gap-3 mb-3 p-2 bg-blue-50 dark:bg-blue-900/30 border border-blue-200 dark:border-blue-700 rounded-lg text-sm">
-          <span className="font-medium text-blue-700 dark:text-blue-300">{ep.selected.size} selected</span>
-          <div className="border-l border-blue-200 dark:border-blue-700 h-5" />
-          <select
-            value={ep.actionTag}
-            onChange={e => ep.setActionTag(e.target.value)}
-            className="px-2 py-1 border border-gray-300 dark:border-gray-600 rounded text-sm dark:bg-gray-700 dark:text-gray-200"
-          >
-            <option value="">Select tag...</option>
-            {ep.tags.map(t => (
-              <option key={t.id} value={t.id}>{t.name}</option>
-            ))}
-          </select>
-          <button
-            onClick={ep.assignTag}
-            disabled={!ep.actionTag || ep.busy}
-            className="px-3 py-1 rounded text-sm font-medium text-white bg-green-600 hover:bg-green-700 disabled:opacity-50"
-          >
-            Assign Tag
-          </button>
-          <button
-            onClick={ep.removeTagFromSelected}
-            disabled={!ep.actionTag || ep.busy}
-            className="px-3 py-1 rounded text-sm font-medium text-red-600 dark:text-red-400 hover:bg-red-50 dark:hover:bg-red-900/30 border border-red-200 dark:border-red-700 disabled:opacity-50"
-          >
-            Remove Tag
-          </button>
-          {ep.hasAnyFilter && ep.total > ep.selected.size && (
-            <>
-              <div className="border-l border-blue-200 dark:border-blue-700 h-5" />
-              <button
-                onClick={ep.assignTagToAll}
-                disabled={!ep.actionTag || ep.busy}
-                className="px-3 py-1 rounded text-sm font-medium text-blue-700 dark:text-blue-300 hover:bg-blue-100 dark:hover:bg-blue-900/30 border border-blue-300 dark:border-blue-700 disabled:opacity-50"
-                title={`Tag all ${ep.total} users matching current filters`}
-              >
-                Tag all {ep.total} matching
-              </button>
-            </>
-          )}
-          <button
-            onClick={() => ep.setSelected(new Set())}
-            className="px-2 py-1 rounded text-xs text-gray-500 dark:text-gray-400 hover:bg-gray-100 dark:hover:bg-gray-700 ml-auto"
-          >
-            Clear selection
-          </button>
-        </div>
+      renderDataCells={(u) => (
+        <>
+          <td className="px-3 py-2 text-gray-600 dark:text-gray-400 text-xs">{u.userPrincipalName}</td>
+          <td className="px-3 py-2 text-gray-600 dark:text-gray-400">{u.department || ''}</td>
+          <td className="px-3 py-2 text-gray-600 dark:text-gray-400">{u.jobTitle || ''}</td>
+        </>
       )}
-
-      {/* Table */}
-      {ep.loading ? (
-        <div className="text-center text-gray-500 dark:text-gray-400 py-12">Loading users...</div>
-      ) : ep.items.length === 0 ? (
-        <div className="text-center text-gray-500 dark:text-gray-400 py-12">
-          {ep.hasAnyFilter ? 'No users match the current filters.' : 'No users found.'}
-        </div>
-      ) : (
-        <div className="border border-gray-200 dark:border-gray-700 rounded-lg overflow-hidden">
-          <table className="w-full text-sm">
-            <thead>
-              <tr className="bg-gray-50 dark:bg-gray-700 border-b border-gray-200 dark:border-gray-700">
-                <th className="w-10 px-3 py-2">
-                  <input
-                    type="checkbox"
-                    checked={ep.allOnPageSelected}
-                    onChange={ep.toggleSelectAll}
-                    className="rounded"
-                  />
-                </th>
-                {TABLE_COLUMNS.map(col => (
-                  <th
-                    key={col.key}
-                    onClick={() => ep.toggleSort(col.key)}
-                    className="text-left px-3 py-2 font-medium text-gray-700 dark:text-gray-300 cursor-pointer select-none hover:bg-gray-100 dark:hover:bg-gray-700"
-                  >
-                    <span className="inline-flex items-center gap-1">
-                      {col.label}
-                      {ep.sortCol === col.key ? (
-                        <span className="text-blue-600 text-[10px]">{ep.sortDir === 'asc' ? '▲' : '▼'}</span>
-                      ) : (
-                        <span className="text-gray-500 dark:text-gray-500 text-[10px]">{'▴'}</span>
-                      )}
-                    </span>
-                  </th>
-                ))}
-                <th className="text-left px-3 py-2 font-medium text-gray-700 dark:text-gray-300">Tags</th>
-              </tr>
-            </thead>
-            <tbody>
-              {ep.sortedItems.map(u => (
-                <tr
-                  key={u.id}
-                  className={`border-b border-gray-100 dark:border-gray-700 hover:bg-gray-50 dark:hover:bg-gray-700/50 cursor-pointer ${
-                    ep.selected.has(u.id) ? 'bg-blue-50 dark:bg-blue-900/30' : ''
-                  }`}
-                  onClick={() => ep.toggleSelect(u.id)}
-                >
-                  <td className="px-3 py-2 text-center" onClick={e => e.stopPropagation()}>
-                    <input
-                      type="checkbox"
-                      checked={ep.selected.has(u.id)}
-                      onChange={() => ep.toggleSelect(u.id)}
-                      className="rounded"
-                    />
-                  </td>
-                  <td className="px-3 py-2 font-medium text-blue-600 hover:text-blue-800 cursor-pointer"
-                    onClick={() => onOpenDetail?.('user', u.id, u.displayName)}>
-                    {u.displayName}{u.deletedAt && <> <DeletedBadge at={u.deletedAt} /></>}
-                  </td>
-                  <td className="px-3 py-2 text-gray-600 dark:text-gray-400 text-xs">{u.userPrincipalName}</td>
-                  <td className="px-3 py-2 text-gray-600 dark:text-gray-400">{u.department || ''}</td>
-                  <td className="px-3 py-2 text-gray-600 dark:text-gray-400">{u.jobTitle || ''}</td>
-                  <td className="px-3 py-2">
-                    <div className="flex flex-wrap gap-1">
-                      {u.tags.map(t => (
-                        <span
-                          key={t.id}
-                          className="inline-block px-1.5 py-0.5 rounded-full text-[10px] font-medium border"
-                          style={{ backgroundColor: t.color + '20', borderColor: t.color, color: t.color }}
-                        >
-                          {t.name}
-                        </span>
-                      ))}
-                    </div>
-                  </td>
-                </tr>
-              ))}
-            </tbody>
-          </table>
-        </div>
-      )}
-
-      {/* Pagination */}
-      {ep.totalPages > 1 && (
-        <div className="flex items-center justify-between mt-3 text-sm text-gray-600 dark:text-gray-400">
-          <span>
-            Showing {ep.page * ep.PAGE_SIZE + 1}&ndash;{Math.min((ep.page + 1) * ep.PAGE_SIZE, ep.total)} of {ep.total.toLocaleString()}
-          </span>
-          <div className="flex items-center gap-2">
-            <button
-              onClick={() => ep.setPage(p => Math.max(0, p - 1))}
-              disabled={ep.page === 0}
-              className="px-3 py-1 rounded border border-gray-300 dark:border-gray-600 hover:bg-gray-50 dark:hover:bg-gray-700/50 disabled:opacity-40"
-            >
-              Prev
-            </button>
-            <span>Page {ep.page + 1} of {ep.totalPages}</span>
-            <button
-              onClick={() => ep.setPage(p => Math.min(ep.totalPages - 1, p + 1))}
-              disabled={ep.page >= ep.totalPages - 1}
-              className="px-3 py-1 rounded border border-gray-300 dark:border-gray-600 hover:bg-gray-50 dark:hover:bg-gray-700/50 disabled:opacity-40"
-            >
-              Next
-            </button>
-          </div>
-        </div>
-      )}
-    </div>
+      searchPlaceholder="Search by name or UPN..."
+      showIncludeDeleted
+      subTabBar={subTabBar}
+      baseFilters={baseFilters}
+      customizeFilterFields={customizeFilterFields}
+      onOpenDetail={onOpenDetail}
+    />
   );
 }
